+++ src/collective/messagesviewlet/utils.py
@@ -1,15 +1,20 @@
 # -*- coding: utf-8 -*-
 
+from Acquisition import aq_parent
 from collective.behavior.talcondition.behavior import ITALCondition
-from collective.messagesviewlet import HAS_PLONE_5
+from collective.messagesviewlet import HAS_PLONE_5_AND_MORE
 from collective.messagesviewlet.message import add_timezone
 from collective.messagesviewlet.message import generate_uid
+from collective.messagesviewlet.messagesconfig import MessagesConfig
 from datetime import datetime
-from DateTime import DateTime
 from plone import api
-from plone.app.layout.navigation.defaultpage import isDefaultPage
+try:
+    from plone.app.layout.navigation.defaultpage import isDefaultPage
+except ImportError as error:
+    from Products.CMFPlone.defaultpage import is_default_page as isDefaultPage
 from plone.app.layout.navigation.interfaces import INavigationRoot
 from plone.app.textfield.value import RichTextValue
+from Products.CMFPlone.interfaces import IPloneSiteRoot
 from six import text_type
 from zope.annotation import IAnnotations
 from zope.component import queryUtility
@@ -30,7 +35,7 @@ def _richtextval(text):
 
 
 def add_message(id, title, text, msg_type='info', can_hide=False, start=datetime.now(), end='', req_roles=[],
-                location='fullsite', tal_condition='', roles_byp_talcond=[], use_local_roles=False, activate=False):
+                location='fullsite', tal_condition='', roles_byp_talcond=[], use_local_roles=False, activate=False, container='default'):
     """
         Add a message in the configuration folder
             msg_type: info, significant, warning
@@ -39,20 +44,21 @@ def add_message(id, title, text, msg_type='info', can_hide=False, start=datetime
             location: fullsite, homepage
     """
     site = api.portal.get()
-    config = site['messages-config']
+    if container == 'default':
+        container = site['messages-config']
     # We pass if id already exists
-    if id in config:
+    if id in container:
         return None
     rich_text = _richtextval(text)
     # Add TZ when using Plone5
-    if HAS_PLONE_5:
+    if HAS_PLONE_5_AND_MORE:
         start = add_timezone(start, force=True)
     try:
         end_date = datetime.strptime(end, '%Y%m%d-%H%M')
         end_date = add_timezone(end_date, force=True)
     except ValueError:
         end_date = None
-    message = api.content.create(container=config, type='Message', id=id, title=title,
+    message = api.content.create(container=container, type='Message', id=id, title=title,
                                  **{'msg_type': msg_type, 'text': rich_text, 'can_hide': can_hide,
                                     'start': start, 'end': end_date, 'required_roles': req_roles,
                                     'location': location, 'hidden_uid': generate_uid(),
@@ -87,7 +93,7 @@ def get_messages_to_show(context, caching=True):
             mb_roles = set(['Anonymous'])
         else:
             mb_roles = set(api.user.get_roles(obj=context))
-        now = DateTime()
+        now = datetime.now()
         brains = catalog.unrestrictedSearchResults(portal_type=['Message'],
                                                    start={'query': now, 'range': 'max'},
                                                    end={'query': now, 'range': 'min'},
@@ -95,10 +101,35 @@ def get_messages_to_show(context, caching=True):
                                                    sort_on='getObjPositionInParent')
         for brain in brains:
             message = brain._unrestrictedGetObject()
+            message_container = aq_parent(message)
+            parent = context
+            if not IPloneSiteRoot.providedBy(context):
+                parent = aq_parent(context)
+
+            if message.location == 'fullsite':
+                nav_root = api.portal.get_navigation_root(context)
+                message_nav_root = api.portal.get_navigation_root(message)
+                if isinstance(message_container, MessagesConfig) or IPloneSiteRoot.providedBy(message):
+                    # message is on site root / config, should always be visible
+                    pass
+                elif message_nav_root.absolute_url() in nav_root.absolute_url():
+                    # message is in same navigation root, should be visible
+                    pass
+                else:
+                    continue
             if message.location == 'homepage':
                 # Test if context is PloneSite or its default page
                 if not INavigationRoot.providedBy(context) and \
                         not isDefaultPage(portal, context):
+                    continue
+            elif message.location == 'justhere':
+                abs_url = context.absolute_url()
+                if not INavigationRoot.providedBy(context) and getattr(parent, "default_page", False) and parent.default_page == context.id:
+                    abs_url = parent.absolute_url()
+                if abs_url != message_container.absolute_url():
+                    continue
+            elif message.location == 'fromhere':
+                if message_container.absolute_url() not in context.absolute_url():
                     continue
             # check in the cookie if message is marked as read
             if message.can_hide:
@@ -113,12 +144,9 @@ def get_messages_to_show(context, caching=True):
             if not ITALCondition(message).evaluate(extra_expr_ctx={'context': context}):
                 continue
             # We check the local roles
-            if message.use_local_roles and \
-               not api.user.is_anonymous() and \
-               'Reader' not in api.user.get_roles(obj=message):
+            if message.use_local_roles and not api.user.is_anonymous() and 'Reader' not in api.user.get_roles(obj=message):
                 continue
             messages.append(message)
-
         if caching:
             cache[key] = messages