+++ CHANGES.rst
@@ -5,7 +5,20 @@ Changelog
 0.24 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Migrate to Plone 6.0.0: remove dexteritytextindexer, use new simplified
+  resources registry, fix styles, 
+  [boulch, laulaz]
+- Add local messages feature. Local messages can be added in any folderish
+  content types and you can choose if / on which levels they display.
+  [boulch]
+- Protect messages-config folder with one-state private workflow
+  [boulch]
+- Add control panel (with messages-config folder link)
+  [boulch]
+- Use JS to show/hide messages when closed, to avoid caching problems (#12)
+  [laulaz]
+- Update / improve translations
+  [boulch, laulaz]
 
 
 0.23 (2020-04-17)